[FIX] Fix execution of tests and benchmarks using memory protection.
RISCV-tests assumes no memory protection at all (at least, my version of it), so
the priviledge code lives near the user code, and it doesn't work unless the
boot address is in the USER memory space.

This adds a little loader that will jump to the first address in user space,
keeping the boot address inside the priviledge (machine) memory space.